Move auto-subscribe logic to another crate
