+++ chains/ethereum/server/src/log_filter.rs
@@ -54,8 +54,8 @@ impl LogFilter {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use hex_literal::hex;
-    use rosetta_config_ethereum::ext::types::Bloom;
+    // use hex_literal::hex;
+    // use rosetta_config_ethereum::ext::types::Bloom;
 
     #[test]
     fn add_remove_works() {
@@ -72,29 +72,32 @@ mod tests {
         assert!(filter.is_empty());
     }
 
-    #[test]
-    fn filter_topics_works() {
-        let mut filter = LogFilter::new();
-        let logs_bloom = Bloom::from(hex!("00000000000000000000000000000000000000000000020200040000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000002000000000000000002000000000000004000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000200000000000000000"));
+    // #[test]
+    // fn filter_topics_works() {
+    //     let mut filter = LogFilter::new();
+    //     let logs_bloom =
+    // Bloom::from(hex!("
+    // 00000000000000000000000000000000000000000000020200040000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000002000000000000000002000000000000004000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000200000000000000000"
+    // ));
 
-        // Empty filter
-        let mut logs = filter.topics_from_bloom(logs_bloom);
-        assert!(logs.next().is_none());
-        drop(logs);
+    //     // Empty filter
+    //     let mut logs = filter.topics_from_bloom(logs_bloom);
+    //     assert!(logs.next().is_none());
+    //     drop(logs);
 
-        let expect_address = Address::from(hex!("97be939b2eb5a462c634414c8134b09ebad04d83"));
-        let expect_topics = [
-            H256(hex!("b7dbf4f78c37528484cb9761beaca968c613f3c6c534b25b1988b912413c68bc")),
-            H256(hex!("fca76ae197bb7f913a92bd1f31cb362d0fdbf27b2cc56d8b9bc22d0d76c58dc8")),
-        ];
-        filter.add(expect_address, expect_topics.into_iter());
+    //     let expect_address = Address::from(hex!("97be939b2eb5a462c634414c8134b09ebad04d83"));
+    //     let expect_topics = [
+    //         H256(hex!("b7dbf4f78c37528484cb9761beaca968c613f3c6c534b25b1988b912413c68bc")),
+    //         H256(hex!("fca76ae197bb7f913a92bd1f31cb362d0fdbf27b2cc56d8b9bc22d0d76c58dc8")),
+    //     ];
+    //     filter.add(expect_address, expect_topics.into_iter());
 
-        let mut logs = filter.topics_from_bloom(logs_bloom);
-        let (address, mut topics) = logs.next().unwrap();
-        assert_eq!(address, expect_address);
-        assert_eq!(topics.next().unwrap(), expect_topics[0]);
-        assert_eq!(topics.next().unwrap(), expect_topics[1]);
-        assert!(logs.next().is_none());
-        assert!(topics.next().is_none());
-    }
+    //     let mut logs = filter.topics_from_bloom(logs_bloom);
+    //     let (address, mut topics) = logs.next().unwrap();
+    //     assert_eq!(address, expect_address);
+    //     assert_eq!(topics.next().unwrap(), expect_topics[0]);
+    //     assert_eq!(topics.next().unwrap(), expect_topics[1]);
+    //     assert!(logs.next().is_none());
+    //     assert!(topics.next().is_none());
+    // }
 }
